Refactor Electron main process and add backend boilerplate
