+++ backend/server.js
@@ -1,0 +1,14 @@
+const express = require('express');
+
+const app = express();
+const PORT = process.env.PORT || 3001;
+
+app.use(express.json());
+
+app.get('/health', (req, res) => {
+  res.json({ status: 'ok' });
+});
+
+app.listen(PORT, () => {
+  console.log(`Backend service listening on port ${PORT}`);
+});
